Create application
Here I load the Tk library and the themed widgets required by Tk.

And I create the GUI of our app. I declare the title of the app, a frame, where the rest of the UI content will be visible by the end, with a nice padding and tell the app to adjust to size changes.

Signed-off-by: Patrick Kontschak <patrick.kontschak@gmail.com>