Create UI elements
I added an input field where the user can enter a number, a button that, when triggered, will calculate the given number (interpreted as feet) to meters and a label that will output the calculated number.

At this point the button doesn't do anything though. I still have to define the function triggered by the button.

Signed-off-by: Patrick Kontschak <patrick.kontschak@gmail.com>