Prettify UI
I added some padding to the grid so that the elements aren't so close to each other. It shouldn't look squished together.

I also added the focus on the input so that the user doesn't have to click in the field before using the app.

And I made it so that if you press 'Return'/'Enter' the calculate function will be triggered just like it would if you pressed the button.

Signed-off-by: Patrick Kontschak <patrick.kontschak@gmail.com>